requirements.txt
numpy
pandas
matplotlib
tensorflow>=2.10
tensorflow-probability
tf-agents
gin-config
opencv-python
protobuf
pillow
pytz
holidays
pint
python-dotenv
bidict
requests
plotly
mediapy
jupyter
